+++ packages/asc-ui/src/components/Typography/Typography.tsx
@@ -2,11 +2,17 @@ import React from 'react'
 import TypographyStyle, {
   Props as TypographyStyleProps,
 } from './TypographyStyle'
+import { deprecatedWarning } from '../../utils'
 
 export type Props = TypographyStyleProps & React.HTMLAttributes<HTMLElement>
 
-const Typography: React.FC<Props> = ({ children, ...otherProps }) => (
-  <TypographyStyle {...otherProps}>{children}</TypographyStyle>
-)
+const Typography: React.FC<Props> = ({ children, ...otherProps }) => {
+  if (otherProps.color) {
+    deprecatedWarning(
+      'You are using the deprecated `color` prop in the `Typography` component. Please use your own custom style rules',
+    )
+  }
+  return <TypographyStyle {...otherProps}>{children}</TypographyStyle>
+}
 
 export default Typography