+++ packages/asc-ui/src/components/Typography/TypographyStyle.ts
@@ -5,12 +5,16 @@ import { Theme } from '../../types'
 export type Props = {
   gutterBottom?: number
   element?: Variant
-  color?: Theme.ColorType
   fontSize?: number
   styleAs?: keyof Theme.TypographyElements
   as?: any
   forwardedAs?: any
   strong?: boolean
+  /**
+   * Deprecated: Use your own custom style rules
+   * @deprecated
+   */
+  color?: Theme.ColorType
 }
 
 export const defaultTypographyStyles = {